Using UInt64/UInt32 in place of IntPtr, removing duplicate util classes

+++ Squalr.Engine.Architecture/Disassembler/IDisassembler.cs
@@ -14,7 +14,7 @@
         /// <param name="isProcess32Bit">Whether or not the assembly is in the context of a 32 bit program.</param>
         /// <param name="baseAddress">The address where the code is rebased.</param>
         /// <returns>A string containing the assembly.</returns>
-        Instruction[] Disassemble(Byte[] bytes, Boolean isProcess32Bit, IntPtr baseAddress);
+        Instruction[] Disassemble(Byte[] bytes, Boolean isProcess32Bit, UInt64 baseAddress);
     }
     //// End class
 }
